Reworked login screen
Sign in button placement has been resolved. To stick to 120 pixels above the bottom of the screen on every device.
The logo image is no longer used from the local machine, it’s used as a url now so we don’t have to worry about changing the picture location